+++ src/app/ChannelUseCasesFactory.java
@@ -4,16 +4,15 @@ import interface_adapter.ViewManagerModel;
 import interface_adapter.go_to_channel.GoToChannelController;
 import interface_adapter.go_to_channel.GoToChannelPresenter;
 import interface_adapter.go_to_channel.GoToChannelViewModel;
-import interface_adapter.go_to_chatl_list.GoToChatListController;
-import interface_adapter.go_to_chatl_list.GoToChatListPresenter;
-import interface_adapter.go_to_chatl_list.GoToChatListViewModel;
+import interface_adapter.go_to_chat_list.GoToChatListController;
+import interface_adapter.go_to_chat_list.GoToChatListPresenter;
+import interface_adapter.go_to_chat_list.GoToChatListViewModel;
 import interface_adapter.refresh_chat_page.RefreshChatPageController;
 import interface_adapter.refresh_chat_page.RefreshChatPagePresenter;
 import interface_adapter.refresh_chat_page.RefreshChatPageViewModel;
 import interface_adapter.send_message.SendMessageController;
 import interface_adapter.send_message.SendMessagePresenter;
 import interface_adapter.send_message.SendMessageViewModel;
-import use_case.go_to_channel.GoToChannelDataAccessInterface;
 import use_case.go_to_channel.GoToChannelInputBoundary;
 import use_case.go_to_channel.GoToChannelInteractor;
 import use_case.go_to_channel.GoToChannelOutputBoundary;
