+++ src/interface_adapter/go_to_chat_list/GoToChatListState.java
@@ -1,4 +1,4 @@
-package interface_adapter.go_to_chatl_list;
+package interface_adapter.go_to_chat_list;
 
 import entity.ChatChannel;
 import entity.User;
