+++ src/view/ChannelView/ChannelView.java
@@ -5,9 +5,9 @@ import interface_adapter.go_to_channel.GoToChannelController;
 import interface_adapter.go_to_channel.GoToChannelState;
 import interface_adapter.go_to_channel.GoToChannelViewModel;
 import kotlin.Pair;
-import interface_adapter.go_to_chatl_list.GoToChatListController;
-import interface_adapter.go_to_chatl_list.GoToChatListState;
-import interface_adapter.go_to_chatl_list.GoToChatListViewModel;
+import interface_adapter.go_to_chat_list.GoToChatListController;
+import interface_adapter.go_to_chat_list.GoToChatListState;
+import interface_adapter.go_to_chat_list.GoToChatListViewModel;
 import interface_adapter.refresh_chat_page.RefreshChatPageController;
 import interface_adapter.refresh_chat_page.RefreshChatPageState;
 import interface_adapter.refresh_chat_page.RefreshChatPageViewModel;
