+++ src/view/ChatListView/ChatListView.java
@@ -4,9 +4,9 @@ import entity.ChatChannel;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.go_to_channel.GoToChannelController;
 import interface_adapter.go_to_channel.GoToChannelViewModel;
-import interface_adapter.go_to_chatl_list.GoToChatListController;
-import interface_adapter.go_to_chatl_list.GoToChatListState;
-import interface_adapter.go_to_chatl_list.GoToChatListViewModel;
+import interface_adapter.go_to_chat_list.GoToChatListController;
+import interface_adapter.go_to_chat_list.GoToChatListState;
+import interface_adapter.go_to_chat_list.GoToChatListViewModel;
 import interface_adapter.go_to_personal_profile.GoToPersonalProfileController;
 import interface_adapter.go_to_personal_profile.GoToPersonalProfileViewModel;
 import interface_adapter.go_to_search.GoToSearchController;
