+++ src/view/PersonalProfileView.java
@@ -1,7 +1,7 @@
 package view;
 
-import interface_adapter.go_to_chatl_list.GoToChatListController;
-import interface_adapter.go_to_chatl_list.GoToChatListViewModel;
+import interface_adapter.go_to_chat_list.GoToChatListController;
+import interface_adapter.go_to_chat_list.GoToChatListViewModel;
 import interface_adapter.go_to_personal_profile.GoToPersonalProfileController;
 import interface_adapter.go_to_personal_profile.GoToPersonalProfileState;
 import interface_adapter.go_to_personal_profile.GoToPersonalProfileViewModel;
