+++ src/view/SearchCourseView.java
@@ -1,10 +1,8 @@
 package view;
 
-import interface_adapter.go_to_chatl_list.GoToChatListController;
-import interface_adapter.go_to_chatl_list.GoToChatListState;
-import interface_adapter.go_to_chatl_list.GoToChatListViewModel;
+import interface_adapter.go_to_chat_list.GoToChatListController;
+import interface_adapter.go_to_chat_list.GoToChatListViewModel;
 import interface_adapter.go_to_personal_profile.GoToPersonalProfileController;
-import interface_adapter.go_to_personal_profile.GoToPersonalProfileState;
 import interface_adapter.go_to_personal_profile.GoToPersonalProfileViewModel;
 import interface_adapter.go_to_search.GoToSearchController;
 import interface_adapter.search_course.SearchCourseViewModel;
